@@ -1,13 +1,11 @@
-import 'package:flutter/material.dart';
-
 class PhotoItem {
   PhotoItem({
-    @required this.title,
-    @required this.thumbUrl,
-    @required this.url,
+    required this.title,
+    required this.thumbUrl,
+    required this.url,
   });
 
-  factory PhotoItem.fromJson({@required Map<String, dynamic> json}) {
+  factory PhotoItem.fromJson({required Map<String, dynamic> json}) {
     return PhotoItem(
       thumbUrl: json['thumbnailUrl'],
       title: json['title'],
